crypto/protobufs/protoserver/pyproject.toml
[tool.poetry]
name = "protoserver"
version = "0.1.0"
description = ""
authors = ["Timofey Bakunin <bakunin.t@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pycryptodome = "^3.11.0"
six = "^1.16.0"
protobuf = "^3.18.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
